Setup compilation with clang (#92)
* Fix compilation errors in game funcs

* Fix compilation errors in CRI funcs

* Add some missed funcs to cri-funcs.json

* Add a basic build action

+++ include/sf33rd/Source/Game/PLS02.h
@@ -7,6 +7,6 @@
 s32 random_16();
 s16 cal_attdir(WORK *wk);
 void add_sp_arts_gauge_init(PLW *wk);
-void add_super_arts_gauge(SA_WORK *wk, s32 ix, s32 asag, u16 mf);
+void add_super_arts_gauge(SA_WORK *wk, s16 ix, s16 asag, u8 mf);
 
 #endif
